+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+## 0.2.0 ()
+- Added text alignment props
+- Added border props
+- Added button component
+
 ## 0.1.0 (31 October 2020)
 - Initial public release